Add a LazyIdentity type
This commit adds a LazyIdentity type. This type can be obtained and
used to initialize and perform a GC mark/sweep in a lazy fashion, i.e.
no identity service will be created unless there's been some other
activity related to it.

The reason this is needed is that it's very unnecessary to start the
identity service unless an apply of the the workflow involves at least
one resource step.